Make draggable handle work for grand(^n)children
Draggable handle works so long as the handle is any descendent of the menu, so I changed this line in my instance and it seems to Just Work. Thank you again for making this, and for giving it away for free. I'm getting a ton of use out of it.